Divide NREPEAT by number of cameras for #92
Two different extensions in the raw data with different ONBAND values are actually part of the same "repeat", so divide by the number of possible ONBAND values, i.e., the number of cameras. This assumes a correctly formed recipe: that the number of repeats is constant over the different wavelengths and ONBAND values.