Add Python executable to PATH when configuring the nightly tests
The nightly tests are failing to find Python when configuring despite getting a Python3_ROOT_DIR that should point CMake at finding it. Hopefully, including Python executable in the PATH will help it find Python.

Currently, the tests run because they don't use scripts that call Python to run, but then "ucomp log" is used to filter logs for warnings and errors and that fails. So the email message doesn't show any warnings or errors that might have occurred during the night.