Handling malformed files automatically for #251
This allows malformed files to be skipped automatically instead of having to be
specified in the epochs file. This is required because the distort sub-command
does an inventory with the NO_SKIP keyword -- not skipping bad files marked out
as to not process in the epochs file. This is required because distortion grid
files themselves are files that would be marked as to not be processed by the
normal pipeline.