Removing data for both cameras for a bad frame for #329
Images look much worse when there is not a simultaneous opposite onband camera image.

This could mean that we completely remove some synoptic files. For example, if there are synoptic files with 2 extensions with identical wavelength, polstate, and onband camera, but one has RCAM as a bad frame and one has TCAM as a bad frame, this would have passed previously. It will now fail.