Flags in epochs file and config file to not verify known bad days for #344
The bad days are marked in the epochs file. There is a config file option
verification/verify_bad_days that can be used to specify whether you want to
verify the marked bad days or not. If set to NO, then you can run the
verification on the entire mission without getting any bad days. If set to YES,
then you can check whether anything has changed on the bad days.